refact(integration-tests): extracted propose-confirm-execute factory and applied style fixes
Addresses review feedback on PR #61.

Style fixes (BitSafe Rust skill):
- Combined std and tracing_subscriber imports under brace syntax.
- Replaced inline 3+ segment paths (std::time::Duration, std::sync::Arc,
  std::sync::Mutex, std::path::Path) with `use` imports.
- Broke up four >100-char URL `format!()` strings in utility_onboarding.rs.
- Dropped single-use `pretty_assertions` dev-dep (stdlib `assert_eq!` suffices).
- Added `set -u` to integration-tests/run.sh and env.sh.

Factory extraction:
- Moved `propose_confirm_execute` and `ProposalCycleCtx` from
  utility_onboarding.rs to a shared `tests/common/governance.rs`.
- Replaced the inline propose-confirm-execute cycles in token_custody.rs
  and generic_vote.rs with factory calls (-291 LoC across the two files).
- The factory now includes the cross-participant ACS visibility gates
  introduced in the previous fix, so every governance cycle
  (utility_onboarding x4, token_custody, generic_vote) is protected from
  the same race.
