+++ tests/common/phases/distribute_dars.rs
@@ -1,3 +1,5 @@
+use std::path::Path;
+
 use anyhow::Context;
 use base64::{Engine, engine::general_purpose::STANDARD as B64};
 use serde_json::{Value, json};
@@ -16,7 +18,7 @@ const DAR_FILES: &[&str] = &[
 pub async fn run(f: &mut Fixture) -> anyhow::Result<()> {
     info!("Phase: distribute_dars");
     let manifest_dir = env!("CARGO_MANIFEST_DIR");
-    let dars_dir = std::path::Path::new(manifest_dir).join("releases/v0/rc3");
+    let dars_dir = Path::new(manifest_dir).join("releases/v0/rc3");
 
     let mut entries = Vec::with_capacity(DAR_FILES.len());
     for filename in DAR_FILES {
